+++ data/ProYapper.txt
@@ -2,4 +2,4 @@ T | 0 | sleep
 E | 0 | splashdown | 2024-08-29 1800 | 2024-08-29 2100
 T | 0 | play
 D | 0 | cs2101 | 2024-08-28 1600
-E | 0 | cca | 2028-08-29 1800 | 2028-08-29 2000
+T | 0 | wtf
