+++ src/main/java/Commands/Command.java
@@ -1,3 +1,8 @@
+package Commands;
+import Main.TaskList;
+import UI.Ui;
+import Storage.Storage;
+
 public abstract class Command {
     public abstract void execute(TaskList taskList, Ui ui, Storage storage);
 }
