+++ src/main/java/Commands/DeadlineCommand.java
@@ -1,30 +1,32 @@
+package Commands;
+
+import UI.Ui;
+import Main.TaskList;
+import Tasks.Task;
+import Tasks.Deadline;
+import Storage.Storage;
+
 import java.time.LocalDateTime;
-import java.time.format.DateTimeParseException;
 
-public class EventCommand extends Command{
+public class DeadlineCommand extends Command {
     private final String taskDescription;
-    private final LocalDateTime startWhen;
-    private final LocalDateTime endWhen;
+    private final LocalDateTime dueWhen;
 
-    public EventCommand(String taskDescription, LocalDateTime startWhen, LocalDateTime endWhen) {
+    public DeadlineCommand(String taskDescription, LocalDateTime dueWhen) {
         this.taskDescription = taskDescription;
-        this.startWhen = startWhen;
-        this.endWhen = endWhen;
+        this.dueWhen = dueWhen;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
         if (this.taskDescription == null || this.taskDescription.trim().isEmpty()) {
             ui.showError("deadline WHAT???");
-        } else if (this.startWhen == null) {
-            ui.showError("WHEN DOES IT START???\nPlease use yyyy-MM-dd HHmm format.");
-        } else if (this.endWhen == null) {
-            ui.showError("WHEN DOES IT END???\nPlease use yyyy-MM-dd HHmm format.");
+        } else if (this.dueWhen == null) {
+            ui.showError("WHEN IS THE DEADLINE???\nPlease use yyyy-MM-dd HHmm format.");
         } else {
-            Task newTask = new Event(this.taskDescription, this.startWhen, this.endWhen);
+            Task newTask = new Deadline(this.taskDescription, this.dueWhen);
             taskList.addTask(newTask);
             int numTasks = taskList.size();
-
             ui.showMessage("Got it. I've added this task:");
             ui.showMessage("  " + newTask.toString());
             ui.showMessage("Now you have " + numTasks + " tasks in the list");
