+++ src/main/java/Commands/ExitCommand.java
@@ -1,3 +1,9 @@
+package Commands;
+
+import Main.TaskList;
+import UI.Ui;
+import Storage.Storage;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
