+++ src/main/java/Main/Parser.java
@@ -1,3 +1,7 @@
+package Main;
+
+import Commands.*;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
