+++ src/main/java/Tasks/Event.java
@@ -1,7 +1,11 @@
+package Tasks;
+
+import Tasks.Task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task{
+public class Event extends Task {
     protected LocalDateTime startWhen;
     protected LocalDateTime endWhen;
