+++ src/main/java/Tasks/ToDo.java
@@ -1,6 +1,8 @@
+package Tasks;
+
 import java.time.format.DateTimeFormatter;
 
-public class ToDo extends Task{
+public class ToDo extends Task {
 
     public ToDo(String description) {
         super(description);
